requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.2
dill==0.2.9
Django==2.1.7
django-appconf==1.0.2
django-axes==4.5.4
django-cors-headers==2.4.0
django-ipware==2.1.0
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
gunicorn==19.9.0
idna==2.8
Markdown==3.0.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
numpy==1.16.1
plac==0.9.6
preshed==2.0.1
psycopg2-binary==2.7.7
PyAudio==0.2.11
PyJWT==1.7.1
pytz==2018.9
regex==2018.1.10
requests==2.21.0
Send2Trash==1.5.0
six==1.10.0
soupsieve==1.8
spacy==2.0.18
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
ujson==1.35
urllib3==1.24.1
WebOb==1.8.5
wrapt==1.10.11
xmltodict==0.12.0
git+https://github.com/AntonOfTheWoods/django-wsgi.git
